Change startup code.
I'm getting errors deploying this config to other computers. It seems
related to missing packages and/or error during bytecompilation.

I have modified the initialization code so that, if the environment
variable EMACS_INSTALL is defined, it will start emacs in
/installation mode/, downloading the required packages and
bytecompiling the config files.

The current implementations is a bit lame but I'll give it a try and
see if it fixes the issue.


git-svn-id: svn://localhost/emacslib/emacslib/trunk@543 a9598837-e506-0410-a967-fbf0066ee963